Adicionando validacoes para o controller de estabelecimentos

+++ TaNoMenu/Controllers/EstablishmentController.cs
@@ -1,8 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
+using TaNoMenu.Exceptions;
 using TaNoMenu.Models.Establishments;
 using TaNoMenu.Repositories;
 
@@ -11,7 +15,7 @@ namespace TaNoMenu.Controllers
     
     [ApiController]
     [Route("[controller]")]
-    public class EstablishmentController : ControllerBase
+    public class EstablishmentController : Controller
     {
         private readonly EstablishmentRepository _establishmentRepository;
 
@@ -22,9 +26,16 @@ namespace TaNoMenu.Controllers
         
         [HttpGet]
         [AllowAnonymous]
-        public List<Establishment> Get()
+        public ActionResult<List<Establishment>> Get()
         {
-            return _establishmentRepository.FindAll().ToList();
+            try
+            {
+                return Ok(_establishmentRepository.FindAll().ToList());
+            }
+            catch (Exception exception)
+            {
+                return HandleError(exception);
+            }
         }
         
         [Route("{establishmentId}")]
@@ -32,7 +43,14 @@ namespace TaNoMenu.Controllers
         [AllowAnonymous]
         public ActionResult Show(int establishmentId)
         {
-            return Ok(_establishmentRepository.FindById(establishmentId));
+            try
+            {
+                return Ok(_establishmentRepository.FindById(establishmentId));
+            }
+            catch (Exception exception)
+            {
+                return HandleError(exception);
+            }
         }
         
         [HttpDelete]
@@ -40,8 +58,19 @@ namespace TaNoMenu.Controllers
         [Authorize]
         public IActionResult Delete(int establishmentId)
         {
-            _establishmentRepository.Remove(establishmentId);
-            return Ok();
+            try
+            {
+                int userId = Int32.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+                if (userId != establishmentId)
+                    return Unauthorized("Sem autorização para realizar esta ação");
+                
+                _establishmentRepository.Remove(establishmentId);
+                return Ok("Estabelecimento deletado com sucesso");
+            }
+            catch (Exception exception)
+            {
+                return HandleError(exception);
+            }
         }
         
         [HttpPut]
@@ -49,17 +78,35 @@ namespace TaNoMenu.Controllers
         [Authorize]
         public IActionResult Put(int establishmentId, [FromBody] Establishment establishment)
         {
-            establishment.Id = establishmentId;
-            _establishmentRepository.Update(establishment);
-            return Ok(establishment);
+            try
+            {
+                int userId = Int32.Parse(User.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+                if (userId != establishmentId)
+                    return Unauthorized("Sem autorização para realizar esta ação");
+                
+                establishment.Id = establishmentId;
+                _establishmentRepository.Update(establishment);
+                return Ok(establishment);
+            }
+            catch (Exception exception)
+            {
+                return HandleError(exception);
+            }
         }
 
         [HttpPost]
         [AllowAnonymous]
         public IActionResult Post([FromBody] Establishment establishment)
         {
-            _establishmentRepository.Add(establishment);
-            return Ok(establishment);
+            try
+            {
+                _establishmentRepository.Add(establishment);
+                return Ok(establishment);
+            }
+            catch (Exception exception)
+            {
+                return HandleError(exception);
+            }
         }
     }
 }
